cmd_parser: treat lines starting with '#' as comments
Just to reduce possible confusion when someone tries to copy & paste the
examples into the command line. That is also the reason why I'm not
bothering with comments on the same line as a command (i.e. at the end
of the command). I don't think it needs to be that fancy.