Change capture statistics output
The idea is to make the output more compact and useful.

There are some weird caveats, that are probably obscure enough that it'd
just be noise trying to explain them to the user:

We have a hardware packet counter, but no byte counter, so I changed
"Bytes transmitted" to "Bytes captured" to imply the difference and
possible disconnect between packet and byte counter. (We could probably
add a hardware byte counter, but transferring it along packet metadata
would be a waste.)

"CRC errors" now strictly refers to incorrect FCS. In theory, you could
argue this should apply also if for example the preamble gets corrupted.

The "Buffer fill" display is for software only, while "overflows" is the
addition of hardware and software overflow events. A user might get
non-0 overflows even if the "Buffer fill" never changes from 0%, which
will seem mysterious. But most time they correlate anyway.

[ED1G-7]