device: IRQ related changes for firmware 6
Newer firmware will manage most aspects of IRQ handling on  the device's
CPU, and the host should not touch these things anymore. But keep it
working for older firmware to avoid inconveniences.

(Using old host software with newer devices should sort-of work, but is
not recommended, as it might miss IRQs.)