nose: protect log FIFO writing with a mutex
The idea behind this FIFO is that logging can be done from any thread,
all while we can pass log output to connected API clients.
Unfortunately, this silently assumed that the FIFO can have multiple
writers, but of course that is not the case. So just protect the writer
side with a mutex, so that only 1 thread can write at a time.

Fixes very rarely occurring crashes (usually an assertion failure in
flush_log()).