nose: wire up firmware autospeed setting
The "speed" command will now use the on-chip speed management facility
on newer firmware. For old firmware, this continues to perform direct
MDIO accesses.

This also supports listing all settings in hw_info (for now there's
really only the speed setting), and a command to reset the device to
defaults.