Allow opening device by serial number
The serial number is unique and stable, so this is much better than the
"device name", which is really just the path/address over which the
device is connected. Using the old "device name" is still supported
though.

Wireshark extcap also displays the serial number now. Though for certain
reasons it still uses the device number internally.

Actually, this was always planned this way, and should be the preferred
method.