cmd_parser: extend some command parsing facilities to option parser
This adds "aliases" to the option parser, so the user can pass symbolic
names as option values instead of magic numbers for certain things. To
be used in a later commit.

Ah, this is what happens when you decide you want it to be simple
instead of fancy, but then you add new requirements yourself. To be
honest, it could have been done without this.