nose: improve logging with \r
A trailing \r is now interpreted specially by the "direct" logging code
(which is used only outside of the nose event loop, for reasons which I
can't comprehend anymore). Code which is logging repetitive progress
messages can use this to make the output nicer.

The flashing code in usb_control.c will make use of this. It happens to
use the "direct" logging path, so it actually works.

Non-direct logging (why did this exist again? I think for the JSON
API...?) will simply treat \r like \n, which, if the convention is
followed, looks reasonable. Though I didn't actually test it.

There are atomics in the code. This is just a concession to the fact
that multiple threads might invoke this code; the existing code didn't
need a lock, and for the new code it would not make that much sense
anyway, because threads could write to stdout in other ways. Relaxed
memory ordering would do fine as well, but I'm too lazy to type it out.