Redo firmware update command line handling
It seems bad for the firmware updating code to just grab a "random"
device and update it. Make it more explicit.

There's now an interactive mode, which lets the user select which device
to update. There's also a mode to mass-update all found devices.

Also add a version check to avoid accidental firmware downgrade. This
can be disabled with yet another option.

This is an incompatible change. Before this change, it wrote the
firmware to the first device it found. A user following the old manual
might be confused, but it's unlikely anyone is going to critically
depend on the old behavior (this is going to be the first firmware
update).

The actual firmware flashing process is the same. The usb_fw_update()
function got trashed and inlined, because I didn't see any advantages in
trying to keep that interface.