Redo packet injector/disruptor
This changes the user interface and the implementation of the inject and
disrupt commands. Although fairly similar to before, this is an
incompatible change. It exposes more features, and the logic to setup
the hardware state machine is now mostly in the host program.

For the sake of weak compatibility with older firmware, half-tested
compatibility code is left in place. So this should mostly still work
with old firmware, but I can't recommend it. For me, the main use of the
compat code is for easier testing with old firmware, not because I want
end users to keep using old firmware. (The old firmware has bugs, so
users should update.)