nose: add support for libreadline
Using rlwrap is fairly annoying, and users may not even be aware that
they should use rlwrap. So why not make use of libreadline?

This is surprisingly cumbersome and broken. libreadline uses globals
everywhere, supports only 1 global user within the process, messes with
signals, has "partial" abstraction for I/O, which is just barely enough
to integrate with our event loop. On the other hand, it somehow works on
win32. I'm really not sure how, but it's a good thing.

I suppose considering the circumstances, the code here is not even that
bad, even if it's a bit of an ifdef soup.

Using libreadline is optional. If Meson does not find it, the behavior
should be as before.

In the future, we could provide advanced features like auto completion.