nose: add a hack to properly terminate in extcap mode on idle ethernet
If Wireshark extcap mode was used, and if there were no packets on the
Ethernet line (e.g. no link up), then stopping Wireshark capture did not
cause "nose" to terminate. The reason is that the only thing that is
done to stop capturing is closing the pipes. Closing a pipe does not
lead concurrent read() calls to return, nor is there any other simple
way to detect it (whatever is the UNIX justification for these dumb
semantics). The only reliable way to test for this is to attempt to
write data to the pipe.

There isn't really a nice way to do this. The extcap control protocol
has no ping command or similar. I decided to send a command that appends
an empty string to the extcap log window. This works.

I also looked at writing an empty pcapng Interface Statistics Block.
This worked and Wireshark simply ignored it, but I suppose it would end
up in the capture file when saving it (didn't check).