nose: fix speed command declaration
The default value was not set, and due to positional initialization, the
help text was accidentally as the default value. Insert the missing
default value initialization.

(This also caused passing NULL to the %s format in printf, which may
have caused a crash on some systems when using the help command; not
sure if that possibly happened on Windows or OSX.)