cmd_parser: add auto-completion helper
This will be used with libreadline to offer auto-completion.

As far as code goes, 100% correct auto-completion is rather intrusive,
because it either needs to duplicate the command line parser exactly, or
be part of it. So this implementation simply supports a subset of the
command line syntax, which isn't too bad, because it's pretty simple.