nose: add auto-completion with libreadline
Offers auto-completion if nose is built with libreadline and the
internal command line is used. I wanted this mostly for device names
(see the following commit), and option names for the overly complicated
"inject" and "disrupt" commands.

readline offers multiple ways to implement this. We are using
rl_completion_entry_function and not rl_attempted_completion_function,
which seems like a better match at first, but it has weird, undocumented
semantics, and appears to require some additional processing (didn't dig
too deep into it).