Add latency tester feature
This is meant for specific setups with 2 ethernet debugger devices. Each
device is connected with the A/B ports of the other device. One of the
two connections has a device under test inserted. One device sends
packets, the other device receives them, and measures the delay between
them. The resulting delay must be the latency introduced by the DUT.

This is implemented by driving the packet injector by a timer on the
sending side, and automatically parsing the packets on the receiving
side.

This basically simplifies a workflow which could be realized with
existing mechanisms.

One implementation detail is that the packet injector is driven by a
timer. (Avoids generating the packets in the firmware, which would be
possible but inflexible, as every change would require updating the
firmware.) Windows just barely supports waiting for 1ms for the win32
API used by the mainloop. Fortunately, Windows 10 doesn't seem to reduce
the timer precision to 10ms or worse. (Observed timer accuracy is still
worse than Linux, waiting and sending the packet takes about 1.2ms on
win32, while Linux sends it with pretty in 1ms.)

Details are in the user manual.