README: make link to docs more prominent
I got confused when I looked at the "Usage instructions" section, and
didn't see the rather important link to the proper docs hosted on
intona.eu. The link is further up in the "Documentation" section, but
it's better to just put it everywhere where the docs are referenced.
Also make the terminology more consistent.