nose: work around UB on pasting text with newer libreadline
It probably can be considered libreadline's bug. For an exhaustive
explanation see the code comments. It crashes reliably with ASAN, and
tends to insert some random additional characters without.

We should probably find something better or implement our own subset of
libreadline. Or run text input on a separate thread in the first place
to avoid having to use libreadline in a way it pretends to support, but
doesn't really. But that would come with other problems.