device: report packet inject slightly better in a meaningless corner case
regs[1] contains a counter that is incremented when packet injection
starts, and when it stops. regs[2] contains the counter of how many
packets still to inject. Thus, if regs[1] is uneven, a packet is being
injected, and the injector shouldn't be reported as inactive.

This is such a small time window, that it doesn't matter, unless the
injector logic is somehow transmitting for an excessively long time.

The | is there because I was too lazy to make it a saturating add, in
case the other value is UINT32_MAX. (This is the "slightly" part in the
commit subject line.)