cmd_parser: don't list raw values for alias-only parameters in help
Alias-only parameters allow only one of the alias strings, and none of
the "raw" values. In fact, it was introduced as a mechanism to avoid
exposing magic integer values to the user. In this case, the raw value
is misleading, useless, and confusing. Don't print it in the help
output.