nose: make --ipc-server accept full path as well
Until now --ipc-server always expected a "symbolic" name, that was
expanded to a full path. I'm not sure why I decided that it should work
this way, but in part it was probably because win32 requires a special
namespace for pipes.

But this might be confusing and restrictive, so allow passing in a full
path as well. If a '/' or '\' character appears in the name, assume it's
a full path.