grabber: keep as much from oversized packets as possible
In theory, the firmware's FIFO and USB protocol can transfer ethernet
frames of unlimited size. The protocol parser in grabber.c uses a fixed
size buffer for simplicity and robustness, which limits the maximum
packet size that can be captured. It is large enough for jumbo frames of
any kind, but can be too short in pathological cases. In this case, the
entire frame was silently discarded (except if --verbosity 2 was used).

To get a chance to notice such cases, pass on the remainder of the
frame, instead of discarding everything.