nose: add gross, untested workaround for sporadic issue
The commit message says it all.

This seems to reduce the probability of an hang at exit, where libusb
seemed not to invoke a callback for a queued transfer.

The exit logic is already messy: it tries to synchronously cancel all
outstanding transfers, all while libusb_cancel_transfer() is
asynchronous. So it has to wait for libusb invoking the callback
(indicating completion of the cancellation). So it cancels all
transfers, and then runs the libusb mainloop to wait for the callbacks.

To make it worse, another thread has to solve a different problem,
signaling that the libusb thread (or rather, the thread on which we call
the libusb mainloop) should exit. Somehow this is hard. Possibly I don't
understand this part of the libusb API correctly. Anyway, it's the mess
it is. The added "break;" in this part of the code is a simplification,
because you don't need to try to send the stop signal once it was sent
(the pthread_join() takes care of waiting for the final deinit of the
libusb thread).

I was able to reproduce this yesterday, and planned to attempt a better
fix, but now I can't reproduce it anymore. Reasons unknown. But when I
could reproduce it (randomly but fairly often), this reduced the
probability of hangs, though not always. The sleeps probably stop the
threads from interfering with each other, possibly removing a livelock,
or reduces the probability of undetected race conditions.

Having an ugly workaround is better than having a known obvious bug sit
around forever, so here we go.