nose: log forced speed mode on link status output
I often get confused when no link is established. Usually, it's when I
force the speed to something incompatible to the connected devices.

Try to make this more convenient by printing the forced speed setting,
when a link change message is printed.

I'm trying to be clever and not print the message in situations where
either the default setting is used ("same" mode aka "autospeed"), or if
the links are at the best possible mode (1000MBit). This way, the
message is never printed if the user never used the "speed" command.

(It would be even better if it reported unsuccessful "autospeed"
negotiations by checking the remote advertisements, but it doesn't do
that.)