usb_io: fix exit handling and exit freezes
This was quite puzzling, and I even suspected libusb bugs. libusb
seemingly didn't run all transfer completion callbacks when cancelling
transfers. But in the end, it was my own stupidity, and the cause and
solution are astonishingly simple: if ctx->interrupt is set to 1, then
libusb_handle_events_completed() will always exit immediately. I just
assumed it would handle all outstanding events, and then just not wait.
But it basically returns immediately. Since ctx->interrupt was never
unset, no events got handled, and no further transfer callbacks
happened. The workaround hack with the timeouts was even more idiotic
than it seemed, and it "helped" only because killing the event handler
was delayed for a bit, making it less likely that there were still
pending transfers.

Fix this by using the fact that libusb_interrupt_event_handler() does
guarantee that libusb_handle_events() returns at least once. (This was
unclear to me until I read more of the libusb source code. It could have
been that the function had certain races, for example when called before
entering the event handler, and might have required using certain
locks.) This greatly simplifies the logic, and we don't need to mess
with stupid locks and other stuff.

usb_cancel_all() was redundant, because usb_ep_remove() must be called
for all EPs before exiting, which cancels all transfers.

Move the code for waiting for all transfers to finish from usb_thread()
to usb_thread_destroy(), because I think it's easier to understand and
to debug. It's not technically necessary.

And now it's time to complain about libusb. At least the Linux backend
is single-threaded and uses a poll() mainloop. (nose uses a separate
thread for libusb mainly because it would be too tricky to integrate our
own mainloop with libusb's on win32.) Yet libusb has tons of different
mutexes and condition variables and stuff to interact with them - what
for? If your library is single-threaded, it's better to have a simple
API and let the user worry about synchronization. Interacting with the
threaded stuff is not simple. libusb_interrupt_event_handler(), the most
useful function, was under-documented. Extremely tricky to use
mechanisms like libusb_handle_events_completed() get too much focus. The
idea of multiple thread to handle events (not concurrently) seems rather
dumb. Requiring the API user to wait for transfer completion after
cancellation is the main reason for all the complexity in usb_io.c, and
with which all API users have to fight with. (Or they ignore it and you
get unreliable libusb programs.) Completion callbacks are a really bad
choice, because they require the user to implement synchronization (a
backend could still choose to call them from random threads), and it
creates API reentrancy issues. There seem to be implementation bugs too:
event_handler_active (struct libusb_context) is read/written under
different two independent mutexes. OK that's all.