nose: correct description of --exit-on default
The default obviously and intentionally makes nose auto-exit if --fifo
was used. But I have to rethink whether this is actually useful to treat
as a special case. The user can stop capture in three ways: using the
capture_stop command, janking out the USB cable, or writing the capture
output to a FIFO and the FIFO is closed. In the first case, the user
could just use the exit command. The second case is obscure. Maybe the
third case makes it reasonable and is why it works like this in the
first place? For now, leave it like it is.