nose: adjust auto-exit behavior (--exit-on, --fifo, etc.)
Normally, nose exits if stdin is closed, because that's what happens if
you for example send ctrl+d to the terminal. (Except with libreadline,
uh, not sure what exactly happens there, let's ignore it.) One of the
reasons this was done was that it shouldn't be possible to accidentally
start nose in a way it just sits in the background doing nothing, with
no way to control it (other than sending UNIX signals). There was also
some code in main() which was supposed to take care of this.

Redo this, and make it dependent on whether there's anything open that
can take commands. If not, it's one of the situations described above,
and exit. This happens in the event loop idle handler, so this is
checked very often but not too often (but at least every time any of
these conditions change).

As a bonus, the "--exit-on capture-end" mode (which is also the default
when using the --fifo option) will now not terminate the program if
capture is stopped and immediately restarted by sending another
capture_start command. This is probably more intuitive.

Hopefully, this covers all cases.