nose: redo exit behavior (--exit-on in particular) yet again
The previous commit was wrong and was pushed too soon. We don't actually
want to keep nose running by default if all input died and capturing is
still active.

The behavior as it should be is:

If nose is started with --wireshark or --fifo, we want nose to exit if
capture stops, because it would be annoying to keep the terminal command
line around after Wireshark or the consumer of --fifo is terminated.
Also, it should be possible to start nose without any input in this
case, so we can say input terminating should not be a reason to
terminate nose. This is the behavior of the "no-capture" mode.

If nose is started without any arguments, we're definitely in an
interactive mode. Stopping capture should not exit nose. Losing control
input should exit, even if capture is active. This is the behavior of
the "no-input" mode.

If nose is started with IPC (--ipc-server or similar), the situation is
actually the same as in interactive mode. An IPC connection of any kind
counts as control input, and as long as it exists, we should not exit.
If it gets lost, we should stop even if capture is going on, because we
have to assume something with the IPC master went wrong, and we should
not hog the USB device any longer.

(One problem with --ipc-server was that it would exit anyway if stdin
was closed, which can happen in scripts. This situation was fixed
earlier.)

At least with --fifo, the user might want the behavior that "interactive
mode" defaults to. For that, "--exit-on no-input" can be used.

The other --exit-on modes are probably not terribly useful in general.
But I feel like "default" is needed to do the right (?) thing,
"no-input"/"no-capture" to force the exact behavior, "never" to make the
auto-magic behavior get out of the way, and "always" is sometimes useful
for scripting.

I considered adding a combined "no-input-no-capture" mode, which would
make nose exit if neither input nor capturing was active. This mode of
behavior is actually removed from this commit (used to print "Nothing to
do and no controlling input."), but I couldn't come up with a reason why
anyone would need this.