grabber: fix UB on obscure error case
When writer_thread terminates with an error during initialization, it
could happen that wbuf was not initialized yet (goto to wbuf cleanup
code before declaration of wbuf). gcc was nice enough to warn me about
this, but only with optimization enabled. This error could never happen,
because it is triggered only by "grabber_filter" initialization, which
no currently existing filter uses.