grabber: tolerate unusual preamble lengths when stripping/setting flags
Ethernet normally has a preamble of 56 bits alternating between 1/0 (7
bytes of 0x55). We ignored this, because it seemed like an artifact from
10 MBit times. But it happens even with Gigabit Ethernet. It shouldn't
be treated as an error.

Although the firmware also hardcodes the preamble length, we handle most
of it in the host tool. The firmware uses it to set flags and to check
the CRC (FCS). Ignore the reported flags, and if the preamble has a
length other than 7, don't rely on the CRC check (pretend the CRC is
OK). Wireshark seems to ignore the FCS flag anyway (it only displays it
deep in the dissection tree).

Fix the preamble check in the host tool. In particular, frames should be
stripped properly with LINKTYPE_ETHERNET now.

The check is kept extremely simple by searching for the 0xD5 byte only.
IEEE 802.3br is ignored for now.