grabber, nose: add feature to strip FCS
Requested. I'd probably enabled this option by default, but the fact
that Wireshark can't check the CRC anymore (and practically ignores our
own CRC flag in the PCAPNG stream) makes me wary.

It's a bit strange that Wireshark doesn't offer this option on its own,
since the PCAPNG headers tell it whether the FCS is present in the
packet data or not.