nose, grabber: add --capture-speed-test option
This can be used to run a speed test.

The option itself actually simply discards capture data coming from USB.
To actually run a speed test, you also need to set --capture-stats,
connect the ethernet debugger with itself, and run the
inject/block_ports commands to generate some kind of ethernet traffic.

The point of this is to test how much bandwidth a certain type of
ethernet traffic generates, and whether the hardware, the USB layer,
your OS, and the basic libusb part of the host tool can handle it. You
can compare the results to full capture to see whether packet parsing
and PCAP output is the bottleneck.