nose, event_loop: fix Wireshark not being able to start capture again
On Windows, stopping Wireshark capture and then starting it again often
didn't work properly. (It depended on your luck, but often it didn't.)

This is because nose.exe seemed to refuse to exit. This in turn is due
to not handling pipe closing properly. It would just sit around and do
nothing. When Wireshark started nose again, couldn't open the USB
device that was locked by the old nose process. The underlying cause
of not exiting is missing handling of PIPE_EVENT_ERROR.

This bug appeared to be Windows specific, because closed pipes just
signal EOF on POSIX. On win32, the code intended to translate it to EOF
as well, but for some reason and in this case, the win32 API is not
signalling ERROR_BROKEN_PIPE, but ERROR_PIPE_NOT_CONNECTED. I'm not
sure what the difference is. (Not bothering with checking for it
explicitly until I find a reason why it'd matter for anything.)

Triggering PIPE_EVENT_ERROR is not strictly a bug. Actually there are
two related issues in the on_extcap_ctrl_in() function causing the
problem:

First, it didn't check PIPE_EVENT_ERROR at all. Can't blame it, who
would check for error flags explicitly? A pipe error should implicitly
act as if the pipe is closed. Adjust event_loop.c accordingly to treat
pipes as fully closed on any errors. The actual close events are
triggered by the other code later. (Technically it allows the reader to
consume the remaining read buffer.)

Second, it even ignored PIPE_EVENT_CLOSED_READ, because it wanted to
read more data to parse the current/next extcap control header.
Obviously calling pipe_peek() and returning if there's not enough data
(continuing to wait for more data) will block it forever if the pipe is
broken. (pipe_peek() has no return value because it doesn't actually
trigger any action, it only checks the buffer.)

Note that at least on Windows, there's a notable delay when stopping
capture. The device LED will continue to blink for 1-2 seconds. If you
try to restart capture during this time, it will error because the
device is still opened by the previous nose process.