device: show slightly better error messages when opening fails
Give a hint that this can be because something is keeping it open,
rather than an actual permissions problem. Sadly it doesn't seem to be
possible to distinguish both cases on Linux, if I understand that right.
(If you try to open a device opened by another process, Windows fails on
liusb_open(), Linux on libusb_claim_interface().)

Also show the exact libusb error by libusb_open(). Even if meaningless
to the user, this can help with support.