nose: extensive extcap workarounds if device is in use
If you start extcap, but the device is either in use, or opening the
device fails for similar reasons, then Wireshark acted strangely.

The following problems have been observed:
- pretending hat capture started and everything is ok
- like the previous problem, but after the actual problem was resolved
  (such as killing the process who kept the device open)
- showing weird error messages when stopping capture (broken pipes,
  or claiming the extcap binary wrote invalid data)
- keeping nose alive (or as zombie and not releasing the zombie)
- not giving any explanation what the hell is going on

Some of these can be observed even with extcap tools distributed by
Wireshark itself. But us enabling extcap control pipes seems to make it
worse.

Try to improve this with a number of workarounds. To be honest, this
should just be fixed upstream in Wireshark (if even possible).

Don't exit immediately when an error is detected. Give Wireshark a
chance to write to the control pipes, because if we don't open the pipes
or close them tool early, Wireshark will freeze or show error messages
that opening failed.

Always open the packet data FIFO. Otherwise Wireshark will misbehave. We
also need to write some (valid) data to it. Just write a header and
nothing else. This is why grabber_get_pcap_dummy_header() and the
previous commit is needed. The normal grabber code needs an USB device
to even start.  This function create a dummy header is pretty weird, but
seems like the better solution, at least for now.

Use the extcap control pipe to get Wireshark to display an error
message. Actually it seems you could this just dump to stderr
(whatever), but currently we don't log to stderr at all, so the log
messages need to be preserved anyway using this extcap_prelog thing.

And as a last thing, wait for 1 second before exiting. While I dislike
timeouts to solve problems that don't need timeouts, this whole
situation seems to be inherently racy. So there's no chance to get it
right.

With this, it seems to work fairly well. Unfortunately, I can still make
it turn nose into a zombie or a detached process (which is still keeping
the device open) by quickly stopping and restarting capture on Linux.
This is not as easy to reproduce and much harder to debug.