nose, event_loop: change event semantics and fix a bug
Before this, the user was supposed to reset the event manually in the
event callback. The comment in the header file has an argument in favor
of these semantics, but to be honest I don't understand it anymore.

The new semantics auto-reset the event, though if another event happens
during execution of the event function (or shortly before that), the
user will get another event. (This _could_ have been an advantage of the
previous semantics, because the reset point was more explicit and the
user had to be aware of it. The event handler still can reset events
manually to avoid redundant events.)

This also fixes a very obvious race condition: the event_set() function
can be called from any thread, but it set has_work without
synchronization. Doing this in the lock wouldn't help, because the
has_work field is supposed to be accessed from the event loop only, and
all other code doesn't access it under the lock.

Technically, both issues are unrelated, but splitting the commit isn't
worth it.