nose: add usb_speed_test command
This is of somewhat questionable usefulness, basically just testing.

It simply sends arbitrary data to a lookback OUT endpoint, and expects
it back from an IN endpoint. It logs errors if there is any unexpected
data. Other than that it just logs the measured speed and number of
packets.

This requires new firmware, because older firmware doesn't have loopback
endpoints.