nose: add re-enumerate command
This is of somewhat questionable usefulness (again).

This sends a special control request to the device, that causes it to
disconnect and reconnect. The host side tries to wait for this and then
attempts to reopen the device.

The sspeed parameter can be set to false to force the USB connection to
USB 2.0 only.

As an additional, extremely obscure feature, the switch_buffers changes
the firmware internal DMA buffers to use a larger buffer for the
loopback endpoint, while reducing the buffers used for ethernet
capturing. (Both can be active at the same time and use a limited DMA
buffer pool.)

All settings are discarded on device power cycles.