Better device/host clock sync (needs new firmware)
The old code basically selected an "approximate" time offset between
device and host clock, while the new code (which relies on new firmware
API) tries to be more exact.

Perhaps the sync code could be improved for old firmware. But there are
a lot of inherent problems (such as timeouts, that flush USB packets
when there hasn't been a new ethernet packet in a while) that make this
seems like a pointless endeavor. Just update the firmware if you need
this.